Remove /using-traps page
This removes the view and the controller for the /using-traps page,
updates the forward and back links that point to /using-traps, removes
the field from the confirm page and removes the logic from the
controller before sending the emails. Apart from updating the email
templates on Notify, this is every reference to the using-traps question
and flow gone.

Issue: Scottish-Natural-Heritage/Licensing#144